training/requirements.txt
beartype==0.15.0
coverage==7.4.0
datasets==2.17.1
einops==0.7.0
filetype==1.2.0
https://github.com/NVIDIA/dllogger/archive/26a0f8f1958de2c0c460925ff6102a4d2486d6cc.zip
inflect==4.1.0
jaxtyping==0.2.25
kenlm==0.2.0
levenshtein_rs==0.1.7
librosa==0.8.0
more-itertools==10.2.0
numpy==1.23
orjson==3.10.6
pandas==1.5.3
pydantic==2.8.2
pytest==7.2.1
qtorch==0.3.0
ruamel.yaml==0.18.6
sentencepiece==0.1.96
snakeviz==2.2.0
soundfile==0.12.1
sox==1.4.1
tensorboard==2.11.2
text_unidecode==1.3
wtpsplit==2.0.6
yappi==1.5.1
zmq==0.0.0
